+++ src/helper.hpp
@@ -24,8 +24,12 @@ SOFTWARE.
 #ifndef SRC_HELPER_HPP_
 #define SRC_HELPER_HPP_
 
-// Includes
+#include <battery.hpp>
+#include <history.hpp>
+#include <led.hpp>
 #include <main.hpp>
+#include <perf.hpp>
+#include <serialdbg.hpp>
 
 #define ERR_FILENAME "/error.log"
 #define ERR_FILENAME2 "/error2.log"
@@ -69,168 +73,7 @@ char* convertFloatToString(float f, char* buf, int dec = 2);
 float reduceFloatPrecision(float f, int dec = 2);
 
 // Logging via serial
-void printTimestamp(Print* _logOutput, int _logLevel);
-void printNewline(Print* _logOutput);
 void printHeap(String prefix = "HELP");
-
-// Classes
-class SerialDebug {
- public:
-  explicit SerialDebug(const uint32_t serialSpeed = 115200L);
-  void begin(Print* p) { getLog()->begin(LOG_LEVEL, p, true); }
-  static Logging* getLog() { return &Log; }
-};
-
-class FloatHistoryLog {
- private:
-  String _fName;
-  float _average = 0;
-  float _runTime[10] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
-  int _count = 0;
-  void save();
-
- public:
-  explicit FloatHistoryLog(String fName);
-  void addEntry(float time);
-  float getAverage() { return _average; }
-};
-
-class BatteryVoltage {
- private:
-  float _batteryLevel = 0;
-
- public:
-  BatteryVoltage();
-  void read();
-  float getVoltage() { return _batteryLevel; }
-};
-
-enum LedColor {
-#if defined(ESP32C3)
-  OFF = 0x000000,
-  BLACK = 0x000000,
-  RED = 0xff0000,
-  GREEN = 0x00ff00,
-  BLUE = 0x0000ff,
-  CYAN = 0x00ffff,
-  PURPLE = 0xff00ff,
-  YELLOW = 0xffff00,
-  WHITE = 0xffffff
-#elif defined(ESP32S3)
-  OFF = 0x000000,
-  BLACK = 0x000000,
-  RED = 0x00ff00,
-  GREEN = 0xff0000,
-  BLUE = 0x0000ff,
-  PURPLE = 0x00ffff,
-  CYAN = 0xff00ff,
-  YELLOW = 0xffff00,
-  WHITE = 0xffffff
-#else
-  OFF = HIGH,
-  BLACK = HIGH,
-  RED = 3,  // TIcker at fast pace
-  GREEN = LOW,
-  BLUE = 2,  // Ticker at slow pace
-  PURPLE = LOW,
-  CYAN = LOW,
-  YELLOW = LOW,
-  WHITE = LOW
-#endif
-};
-
-void ledOn(LedColor l = LedColor::WHITE);
-void ledOff();
-
-#if defined(COLLECT_PERFDATA)
-
-class PerfLogging {
- private:
-  struct PerfEntry {
-    uint32_t start;   // millis()
-    uint32_t end;     // millis()
-    uint32_t max;     // max time in ms
-    const char* key;  // measurement
-
-    PerfEntry* next;  // Next in the linked list
-
-    float mA;  // Power consumption
-    float V;   // Power consumption
-  };
-
-  PerfEntry* first = 0;
-  bool measurePower = false;
-
-  PerfEntry* find(const char* k) {
-    if (first == 0) return 0;
-
-    PerfEntry* pe = first;
-
-    while (strcmp(k, pe->key) != 0) {
-      if (pe->next == 0) return 0;
-      pe = pe->next;
-    }
-    return pe;
-  }
-
-  PerfEntry* add(const char* k) {
-    if (first == 0) {
-      first = new PerfEntry();
-      first->key = k;
-      first->next = 0;
-      first->max = 0;
-      return first;
-    }
-
-    PerfEntry* pe = first;
-
-    while (strcmp(k, pe->key) != 0) {
-      if (pe->next == 0) {
-        pe->next = new PerfEntry();
-        pe->next->key = k;
-        pe->next->max = 0;
-        pe->next->next = 0;
-        return pe->next;
-      }
-
-      pe = pe->next;
-    }
-
-    return pe;
-  }
-
- public:
-  void clear();
-  void start(const char* key);
-  void stop(const char* key);
-  void print();
-  void pushInflux();
-};
-
-extern PerfLogging myPerfLogging;
-
-// Use these to collect performance data from various parts of the code
-#define LOG_PERF_START(s) myPerfLogging.start(s)
-#define LOG_PERF_STOP(s) myPerfLogging.stop(s)
-// #define LOG_PERF_PRINT() myPerfLogging.print()
-#define LOG_PERF_PRINT()
-#define LOG_PERF_CLEAR() myPerfLogging.clear()
-#define LOG_PERF_PUSH() myPerfLogging.pushInflux()
-
-#else
-
-// These will disable the performance collection function
-#define LOG_PERF_START(s)
-#define LOG_PERF_STOP(s)
-#define LOG_PERF_PRINT()
-#define LOG_PERF_CLEAR()
-#define LOG_PERF_PUSH()
-
-#endif  // COLLECT_PERFDATA
-
-// Global instance created
-extern SerialDebug mySerial;
-extern BatteryVoltage myBatteryVoltage;
 
 #endif  // SRC_HELPER_HPP_